+++ modules/AWSPowerShell/Cmdlets/DevOpsGuru/Basic/Add-DGURUNotificationChannel-Cmdlet.cs
@@ -35,9 +35,14 @@ namespace Amazon.PowerShell.Cmdlets.DGURU
     /// <para>
     /// If you use an Amazon SNS topic in another account, you must attach a policy to it
     /// that grants DevOps Guru permission to it notifications. DevOps Guru adds the required
-    /// policy on your behalf to send notifications using Amazon SNS in your account. For
-    /// more information, see <a href="https://docs.aws.amazon.com/devops-guru/latest/userguide/sns-required-permissions.html">Permissions
+    /// policy on your behalf to send notifications using Amazon SNS in your account. DevOps
+    /// Guru only supports standard SNS topics. For more information, see <a href="https://docs.aws.amazon.com/devops-guru/latest/userguide/sns-required-permissions.html">Permissions
     /// for cross account Amazon SNS topics</a>.
+    /// </para><para>
+    /// If you use an Amazon SNS topic in another account, you must attach a policy to it
+    /// that grants DevOps Guru permission to it notifications. DevOps Guru adds the required
+    /// policy on your behalf to send notifications using Amazon SNS in your account. For
+    /// more information, see Permissions for cross account Amazon SNS topics.
     /// </para><para>
     /// If you use an Amazon SNS topic that is encrypted by an Amazon Web Services Key Management
     /// Service customer-managed key (CMK), then you must add permissions to the CMK. For
@@ -61,9 +66,12 @@ namespace Amazon.PowerShell.Cmdlets.DGURU
         /// <para> Information about a notification channel configured in DevOps Guru to send notifications
         /// when insights are created. </para><para>If you use an Amazon SNS topic in another account, you must attach a policy to it
         /// that grants DevOps Guru permission to it notifications. DevOps Guru adds the required
+        /// policy on your behalf to send notifications using Amazon SNS in your account. DevOps
+        /// Guru only supports standard SNS topics. For more information, see <a href="https://docs.aws.amazon.com/devops-guru/latest/userguide/sns-required-permissions.html">Permissions
+        /// for cross account Amazon SNS topics</a>.</para><para>If you use an Amazon SNS topic in another account, you must attach a policy to it
+        /// that grants DevOps Guru permission to it notifications. DevOps Guru adds the required
         /// policy on your behalf to send notifications using Amazon SNS in your account. For
-        /// more information, see <a href="https://docs.aws.amazon.com/devops-guru/latest/userguide/sns-required-permissions.html">Permissions
-        /// for cross account Amazon SNS topics</a>.</para><para>If you use an Amazon SNS topic that is encrypted by an Amazon Web Services Key Management
+        /// more information, see Permissions for cross account Amazon SNS topics.</para><para>If you use an Amazon SNS topic that is encrypted by an Amazon Web Services Key Management
         /// Service customer-managed key (CMK), then you must add permissions to the CMK. For
         /// more information, see <a href="https://docs.aws.amazon.com/devops-guru/latest/userguide/sns-kms-permissions.html">Permissions
         /// for Amazon Web Services KMS–encrypted Amazon SNS topics</a>.</para>