+++ modules/AWSPowerShell/Cmdlets/DevOpsGuru/Basic/Get-DGURUEventSourcesConfig-Cmdlet.cs
@@ -28,8 +28,10 @@ using Amazon.DevOpsGuru.Model;
 namespace Amazon.PowerShell.Cmdlets.DGURU
 {
     /// <summary>
-    /// This operation lists details about a DevOps Guru event source that is shared with
-    /// your  account.
+    /// Returns the integration status of services that are integrated with DevOps Guru as
+    /// Consumer via EventBridge. The one service that can be integrated with DevOps Guru
+    /// is Amazon CodeGuru Profiler, which can produce proactive recommendations which can
+    /// be stored and viewed in DevOps Guru.
     /// </summary>
     [Cmdlet("Get", "DGURUEventSourcesConfig")]
     [OutputType("Amazon.DevOpsGuru.Model.EventSourcesConfig")]