+++ modules/AWSPowerShell/Cmdlets/ElasticFileSystem/Basic/New-EFSTag-Cmdlet.cs
@@ -29,8 +29,8 @@ namespace Amazon.PowerShell.Cmdlets.EFS
 {
     /// <summary>
     /// <note><para>
-    /// DEPRECATED - CreateTags is deprecated and not maintained. Please use the API action
-    /// to create tags for EFS resources.
+    /// DEPRECATED - <code>CreateTags</code> is deprecated and not maintained. To create tags
+    /// for EFS resources, use the API action.
     /// </para></note><para>
     /// Creates or overwrites tags associated with a file system. Each tag is a key-value
     /// pair. If a tag key specified in the request already exists on the file system, this